:monocle_face: try to use NJ trees to cluster breeds
This is a proof of concept to see if we can use NJ trees to cluster breeds.
The idea is to use the genetic distance matrix to build a NJ tree and then use
the tree to cluster the breeds. The tree is built using the `Bio.Phylo` module
from the `biopython` package. The clustering is done by evaluating
the total distance of a internal node. The suggest breed annotation for the
cluster is the breed which occurs the most in the cluster.